refactor: Racing 클래스로 인한 메서드 분리
Racing 객체를 만든 후 관심사를 분리시켰음
